feat(extension): support inline type extensions
This PR makes it possible for extension authors to inline their
type extensions into the extension body. We achieve this with a new
explicit typeHooks field in the constructor input as well as a new
phantom property on builder property to supply types along with a helper
sub-constructor for hiding this nested phantom property.